pyproject.toml
[tool.poetry]
name = "resume-analyzer"
version = "0.1.0"
description = ""
authors = ["Yuri"]

[tool.poetry.dependencies]
python = "^3.10"
google-api-python-client = "^2.147.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
